\name{SEMdir-package}
\alias{SEMdir-package}
\alias{SEMdir}
\docType{package}
\title{
\packageTitle{SEMdir}
}
\description{
\packageDescription{SEMdir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SEMdir}
\packageIndices{SEMdir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SEMdir}

Maintainer: \packageMaintainer{SEMdir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
